+++ index.d.ts
@@ -37,7 +37,7 @@ export {SetReturnType} from './source/set-return-type';
 export {Asyncify} from './source/asyncify';
 export {Simplify} from './source/simplify';
 export {Jsonify} from './source/jsonify';
-export {Finite, Integer, Natural, Negative, NegativeInfinity, NegativeInteger, Positive, PositiveInfinity, PositiveInteger} from './source/numeric-range';
+export {Finite, Integer, Natural, Negative, NegativeInfinity, NegativeInteger, Positive, PositiveInfinity, PositiveInteger} from './source/numeric';
 
 // Template literal types
 export {CamelCase} from './source/camel-case';
